test(server): keep file-backed CORS coverage in integration
